big fix
fixed a problem that had me stuck for ages. zombie prefab had the same coordinates as the tile it should have been spawning on except they were in different places. Fixed by changing the coordinates of the game object parent of all the tiles so that it was at 0,0,0 meaning that the local coordinates (coordinates compared to location of parent)  of the tiles were the same as it's global coordinates. Zombie prefab now spawns on correct tiles and moves right to left at a fixed speed. Added a button to take the user to settings and a slider for volume is in progress as there is no code done to make it do anything. Background added for main menu, settings and level selection. to add contrast. Zombie prefab added.