placing plants
created 4 buttons that detect when clicked and register the active button. this corresponds to a position in an array filled by prefabs of the 4 plants and when clicked on the grid, will place said prefab and the the position in the array is set to -1 so repeated pressing cannot occur. scripts created to select a plant as the active one. another script to place a plant. another script to create the array and set the selected index as -1. created an instance = this so that other scripts can more easily reference anything in this array. current problem now is that plants are spawning at same coordinates no matter what.  screen position of mouse is an arbituary number but world position seems to stay at coordinates 4.5, -3.25, 0 no matter what. Problem must be changed before continuing with any more of the game logic.