fixed plants spawning
Plants when clicked on the icon and then clicked on grid did spawn on the same place. This was fixed by changing my main camera's projection to perspective which is normally for 3D games to orthographic which is normally for 2D games. Plants now place where my mouse is. next step is to ensure they are the correct size as they are varying right now