grid snapping
Created code so that any plants placed on the grid would snap to the centre of this. achieved by creating arrays of grid borders and finding inequality of the 2 x values the mouse pos lays between and 2 y values the mouse pos lay between and finding the midpoin.t didn't work initially because they were being placed on the tile above where they should. this is due to my y borders array going in descending in stead of ascending order. i fixed this by reducing minY by 1 instead of increasing it by 1 but i could have also reversed the order of my array.