prevented plants placing on top of eachother
Have improved the placeplant code so that I have decomposed my subroutine to place it. so that creating the central coordinates and creating the plant are separate subroutines. I also created a subroutine to check whether a tile was full. i did this by creating an array filled with vector3s. and plants that are placed, their central coordinates are added in this array of 40 spaces(40 tiles) and the free space pointer corresponding to the next place to add a vector is increased by 1, unless of course it is full, so I have implemented this with a simple if statement. Currently have a problem that no plants are able to be placed on the square centered at 0,0,0 which may be due to the way I have calculated the central coordinates with minX + 1, minY - 1. May change this by calculating an average of the min and max values of x and y