sun generation & falling
created two scripts, one which generates a sun icon at a random x coordinate on the map at a set y level which is attached to an empty game object. another script that is attached to the sun icon prefab which causes the sun object to move down the screen and stop at a random y coordinate. a problem is that the icon used to show the users amount of sun is also part of this prefab so in future will need to just use a different sprite with the same image attached to display this part of the UI so it is not part of the actual game