started to code clicking sun and incrementing total sun
Have fixed the problem of the sun image falling by doing what i said in my last commit. created a new script called collect sun that should when the mouse is clicked on the sun object go to display sun and increment the sun variable there by 50. I did use AI with the linking of the two scripts on line 5 and 14 but this still doesn't seem to work when clicking. i have a collider so I believe it should but i need to look into this closer at another point in time. Also changed the size of an array in spawnZombie from 1-7 to 1-6 because i only have 5 lanes, meaning the zombies were trying to spawn at an index of the gameobject array that didn't exist.