zombies stop when colliding with plant
This tool longer than expected. I created some code and gave my plants and zombie prefabs box colliders and a rigid body to my zombie too but the zombie didn't stop when colliding with the plant. I then tested this in a separate scene with just 1 plant and a zombie with just the basics needed for the 2 to collide and it still didn't work. after some time i looked it up and found someone with the same problem and the solution was to tick the checkbox "use full kinematic contacts". I have now ran into a new issue that some zombies spawn and don't seem to move at all. I can only assume this means that have been instantiated and for some reason have inCollision set to true so do not move. I think to fix this i may try using a void awake method and setting inCollision to false.