sunflower game logic done
moved the method used by the sunflower to a separate script as it seemed to be running every single frame even without a sunflower spawning. I changed timer from == 0 to <= 0 as it never actually reaches 0 or it does for such a short amount of time that the game cannot recognise it. Disabled my whole prefab folder/empty object because i had a sunflower acting as the template off screen but it was still spawning suns which was causing my game to lag quite a lot which is now fixed